requirements.txt
streamlit
openai
supabase-py
pdfkit
